Added alot of stuff
