+++ alert_state_machine/Services/UTMLiveService.cs
@@ -17,9 +17,9 @@ namespace alert_state_machine.Services
     {
         private readonly PureSocketClusterOptions opts;
         private readonly PureSocketClusterSocket utm;
-        private IMapper _mapper;
+        
 
-        public UTMLiveService(IMapper mapper)
+        public UTMLiveService()
         {
             this.opts = new PureSocketClusterOptions
             {
@@ -28,50 +28,17 @@ namespace alert_state_machine.Services
                 DebugMode = true
             };
             this.utm = new PureSocketClusterSocket("wss://healthdrone.unifly.tech/socketcluster/", opts);
-
-            _mapper = mapper;
         }
 
-        public async Task Connect(string token)
+        public async Task Connect(string token, Action<object, string, object> action)
         {
             this.utm.SetAuthToken(token);
             var channel = await this.utm.CreateChannel("adsb").SubscribeAsync();
-            channel.OnMessage(OnMessage);
+            channel.OnMessage((sender, name, data) => action(sender, name, data));
             await this.utm.ConnectAsync();
         }
 
-        private void OnMessage(object sender, string name, object data)
-        {
-            //Console.WriteLine($"{((PureSocketClusterSocket)sender).InstanceName} {name} : {data} \r\n", ConsoleColor.Green);
-            //Console.WriteLine(data);
 
-            var test = (Dictionary<string, object>)(data);
-            var testa = (IEnumerable<object>)test.Values.FirstOrDefault();
-            if (test.Values.ElementAt(1).ToString() == "ADSB_TRACK")
-                return;
-
-            var result = JsonConvert.SerializeObject(DictionaryToObject(testa), Formatting.Indented);
-            var obj = JsonConvert.DeserializeObject<Models.Message>(result);        
-
-            Console.WriteLine(obj);
-        }
-
-        private static dynamic DictionaryToObject(IEnumerable<object> dictionary)
-        {
-            var expandoObj = new ExpandoObject();
-            var expandoObjCollection = (ICollection<KeyValuePair<String, Object>>)expandoObj;
-
-            foreach (var keyValuePair in dictionary)
-            {
-                var test = (Dictionary<string, object>)keyValuePair;
-                foreach (var v in test)
-                {
-                    expandoObjCollection.Add(v);
-                }
-            }
-            dynamic eoDynamic = expandoObj;
-            return eoDynamic;
-        }
     }
 
 }
