+++ alert_state_machine/appsettings.json
@@ -6,8 +6,8 @@
     "windSpeed": 20
   },
   "Redis": {
-    "Host": "",
-    "Port": ""
+    "Host": "localhost",
+    "Port": "6379"
   },
   "UTM": {
     "clientid": "sduHealthDroneConnect",
